--- packages/web3/src/utils/group.ts
@@ -1,38 +1,0 @@
-/*
-Copyright 2018 - 2022 The Alephium Authors
-This file is part of the alephium project.
-
-The library is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-The library is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License
-along with the library. If not, see <http://www.gnu.org/licenses/>.
-*/
-
-import { TOTAL_NUMBER_OF_GROUPS } from '../constants'
-import djb2 from './djb2'
-
-export function groupFromBytes(bytes: Uint8Array): number {
-  const hint = djb2(bytes) | 1
-  return groupFromHint(hint)
-}
-
-export function groupFromHint(hint: number): number {
-  const hash = xorByte(hint)
-  return hash % TOTAL_NUMBER_OF_GROUPS
-}
-
-export function xorByte(intValue: number): number {
-  const byte0 = (intValue >> 24) & 0xff
-  const byte1 = (intValue >> 16) & 0xff
-  const byte2 = (intValue >> 8) & 0xff
-  const byte3 = intValue & 0xff
-  return (byte0 ^ byte1 ^ byte2 ^ byte3) & 0xff
-}
