+++ packages/web3/src/utils/index.ts
@@ -23,4 +23,3 @@ export * from './utils'
 export * from './subscription'
 export * from './sign'
 export * from './number'
-export * from './group'
